+++ src/wcbot/commandhandlers.go
@@ -67,12 +67,14 @@ func (bot *Bot) handleCommandStatus(ctx *BotContext, command, args string) error
 	return bot.Send(ctx, "whisper", "markdown", fmt.Sprintf("*%v* Nodes currently connected.", len(bot.skyMgrMonitor.connectedNodes)))
 }
 
+/*
 // Handler for heartbeat command
 func (bot *Bot) handleCommandHeartBeat(ctx *BotContext, command, args string) error {
 	log.Debug("Handle command /heartbeat")
 	go bot.botHeartbeatLoop(ctx)
 	return bot.Send(ctx, "whisper", "markdown", fmt.Sprintf(msgHeartBeatStarted, bot.config.Monitor.HeartbeatIntMin))
 }
+*/
 
 func (bot *Bot) handleDirectMessageFallback(ctx *BotContext, text string) (bool, error) {
 	errmsg := fmt.Sprintf("Sorry, I only take commands. '%s' is not a command.\n\n%s", text, msgHelpShort)
@@ -90,6 +92,7 @@ func (bot *Bot) AddGroupMessageHandler(handler MessageHandler) {
 	bot.groupMessageHandlers = append(bot.groupMessageHandlers, handler)
 }
 
+/*
 // Bot Heartbeat Loop
 func (bot *Bot) botHeartbeatLoop(ctx *BotContext) {
 	ticker := time.NewTicker(time.Minute * bot.config.Monitor.HeartbeatIntMin)
@@ -102,16 +105,28 @@ func (bot *Bot) botHeartbeatLoop(ctx *BotContext) {
 		}
 	}
 }
+*/
 
-// monitorEventLoop monitors for event messages from the SkyMgrMonitor (when running)
+// monitorEventLoop monitors for event messages from the SkyMgrMonitor (when running).
+// Its also responsible for managing the Heartbeat (if configured)
 func (bot *Bot) monitorEventLoop(runctx context.Context, botctx *BotContext, statusMsgChan <-chan string) {
+	tickerHB := time.NewTicker(bot.config.Monitor.HeartbeatIntMin)
 
 	for {
 		select {
+		// Monitor Status Message
 		case msg := <-statusMsgChan:
+			log.Debugf("Bot.monitorEventLoop: Status event: %s", msg)
 			bot.Send(botctx, "whisper", "markdown", msg)
+
+		// Heartbeat ticker event
+		case <-tickerHB.C:
+			log.Debug("Bot.monitorEventLoop - Heartbeat event")
+			bot.Send(botctx, "whisper", "markdown", fmt.Sprintf("%s\n\n*Connected Nodes:* %v", msgHeartbeat, len(bot.skyMgrMonitor.connectedNodes)))
+
+		// Context has been cancelled. Shutdown
 		case <-runctx.Done():
-			log.Debugln("Bot.monitorEventLoop - Done Event.")
+			log.Debugln("Bot.monitorEventLoop - Done event.")
 			return
 		}
 	}